% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_endpoint_functions.R
\name{echo_back}
\alias{echo_back}
\title{Echo back a message}
\usage{
echo_back(msg = "")
}
\arguments{
\item{msg}{The message to echo}
}
\value{
a list
}
\description{
Echo back a message
}
\examples{
echo_back(msg = "I am a message and no one will stop me!")
}
