% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_filters.R
\name{log_request}
\alias{log_request}
\title{Simple request logger}
\usage{
log_request(req, res)
}
\arguments{
\item{req}{plumber request environment}

\item{res}{plumber response environment}
}
\description{
Simple request logger
}
