% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_endpoint_functions.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot a histogram of random numbers}
\usage{
plot_histogram()
}
\value{
an image of a histogram
}
\description{
Plot a histogram of random numbers
}
\examples{
plot_histogram()
}
