% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_entrypoint.R
\name{run_api}
\alias{run_api}
\title{Run the plungrexample API}
\usage{
run_api(.host = "0.0.0.0", .port = 8000, ...)
}
\arguments{
\item{.host}{IP of the host, Default: 0.0.0.0}

\item{.port}{Which port to use for the API, Default: 8000}

\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run the plungrexample API
}
\examples{
\dontrun{
run_api()
run_api(.host = "127.0.0.1", port = 4242)}
}
