% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber_endpoint_functions.R
\name{sum_numbers}
\alias{sum_numbers}
\title{Sum two numbers}
\usage{
sum_numbers(a, b)
}
\arguments{
\item{a}{The first number to add}

\item{b}{The second number to add}
}
\value{
a number
}
\description{
Sum two numbers
}
\examples{
sum_numbers(808 + 80)
}
